+++ src/index.js
@@ -31,7 +31,6 @@
  */
 import React from 'react'
 import ReactDOM from 'react-dom'
-import './index.css'
 import App from './pages/App'
 import * as serviceWorker from './serviceWorker'